// src/stadium/interface.ts
export interface IStadium {
    stadiumId: number;
    stadiumName: string;
    stadiumCity: string;
    stadiumState: string;
    stadiumLongitude: string;
    stadiumLatitude: string;
    stadiumAddress: string;    
    createdDate: Date;
    createdBy: number;
    updatedDateTime: Date;
    updatedBy: number;
}